Fixed broken pdf generation code
