+++ Services/CO.CDP.DataSharing.WebApi/Model/SharedSupplierInformation.cs
@@ -4,5 +4,5 @@ public record SharedSupplierInformation
 {
     public required Guid OrganisationId { get; init; }
     public required BasicInformation BasicInformation { get; init; }
-    public required Task<List<ConnectedPersonInformation>> ConnectedPersonInformation { get; init; }
+    public required List<ConnectedPersonInformation> ConnectedPersonInformation { get; init; }
 }
